worker/pyproject.toml
[tool.poetry]
name = "worker"
version = "0.1.0"
description = ""
authors = ["Karn Wong <karn@karnwong.me>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.112.0"
python-dotenv = "^1.0.1"
uvicorn = "^0.30.5"

[tool.poetry.group.tests.dependencies]
mypy = "^1.11.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
